requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.31
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.0
matplotlib==3.6.3
requests==2.28.1
